Use Long as type for @Version properties [ACC-1496]

+++ contentgrid-spring-data-rest/src/test/java/com/contentgrid/spring/test/matchers/ETagHeaderMatcher.java
@@ -12,7 +12,7 @@ import org.springframework.test.web.servlet.ResultMatcher;
 @RequiredArgsConstructor(access = AccessLevel.PACKAGE)
 public class ETagHeaderMatcher {
 
-    public static ETag toETag(int version) {
+    public static ETag toETag(long version) {
         return ETag.from(String.valueOf(version));
     }
